Move twelf execution slices to idle portion of main event loop (#35)
Resolves #34 and #36. When I run a big development like
`classical-s5.elf`, I can still manipulate the user interface while it
runs for a few seconds. This seems to be evidence that it's working.
Cancellation, implemented as a menu command with conventional shortcut
⌘-. seems to work ok. I can ⌘-E and ⌘-. in quick succession on
`classical-s5.elf`, and it only gets partway through checking, as
evidenced by the output in the output window.